Fix macos build needing coreutils
Installing coreutils is currently required to build under macos. But
it's only a couple of details that need fixing to make the script
croos-platform. We consider it's worth fixing that so that macos users
can build the package straight away, given than installing coreutils on
macosx can be not desirable for some people.